Bugfix: recovery/config
When a board with no parition table is started the configuration can't
initialize.

However, it should when the default parition table is installed. This
fix adds a call to config_init() when the default partition table is
installed.

Bugfix: recovery/config

When a board with no parition table is started the configuration can't
initialize.

However, it should when the default parition table is installed. This
fix adds a call to config_init() when the default partition table is
installed.

USB Reset wip

Bugfix imx/ehci:

USB could not re-enumerate when the connector was disconnected and then
re-connected. A bit in the EHCI_ENDPTCTRL -registers needed to be set
for every configuration event.

"Whenever a configuration event is received for this Endpoint, software
must write a one to this bit in order
to synchronize the data PID's between the host and device."