Bugfix: imx/usdhc
A change was introduced in 'usdhc_emmc_wait_for_de' to make the function
return PB_OK whenever there are no transfers ongoing. The idea was to
allow other code to call this function to ensure that there are no
ongoing transfers, the previous version would fail/timeout if it was
called when there was no ongoing transfer.

With this change a status flag was introduced
(usdhc_device->transfer_in_progress) and this flag should be set
whenever a transfer is queued. This was to for the
usdhc_emmc_xfer_blocks but not for usdhc_emmc_read_extcsd which caused
problems during initialization. The effect was that it was not possible
to reliably read the extcsd because the function did not wait until the
extcst data was available.

The fix was to set the dev->transfer_in_progress to true before
usdhc_emmc_wait_for_de is called in usdhc_emmc_read_extcsd.