Update bpak
This update adds out-of-bounds checks of meta data when the
bpak_valid_header function is called.

This commit also adds additional checks when the signature data is
copied to an internal buffer to protect agains writing out side of the
buffer.

Additional checks are also performed before metadata pointers are
de-referenced.