plat: imx8x: watchdog: code alignment with ATF
The watchdog initialisation code would configure the watchdog for all
partitions on the board, this had a tendency of hanging the system. When
the watchdog is configured from Linux via ATF, ATF gets the current
partition using sc_rm_get_partition and configures the watchdog for it.
Aligning the code with ATF makes the watchdog reliably restart the
board.