plat: imx8x: make revocation command use sc_seco_commit
On the iMX8, key revocation is performed by committing a revocation mask
configured in the boot image header.

The revoke command is rewritten to ignore the incoming argument and
perform the commit of what's in the boot header. It compares the
revocation fuse before and after the commit operation to detect whether
any revocation was performed.