usb: Pass the correct pointer to pb_usb_close_handle()
`h` of struct pb_usb_private is a libusb_device_handle*, which was
passed to pb_usb_close_handle() which takes a struct pb_usb_private*
instead of the private itself. This would cause a segfault on cleanup.

The compiler doesn't complain about this, so there's a compilation flag
issues as well somewhere.