python: Change lifetime handling of Session objects
Python requires _all_ type objects to implement tp_dealloc, not having
one set is equivalent to tagging the object as a singleton that cannot
be removed. Newer python versions also have a tp_finalize that should be
used for "complex" finalization, however it will never be called if
tp_dealloc is not set.

To allow for re-creation of Session() objects within a python session
(that is, without restarting the interpreter) change the previous
tp_finalize to be tp_dealloc instead. The finalization done is quite
limited to no need to split it up.