command/bio/mmc: New way of dealing with partition tables
Sometimes it might be necessary to resize partitions. Historically this
has been done with the 'resize' command. With this commit the resize
command is deprecated.

Changing partition sizes has been a two step operation, first the
default partition table is installed with 'part --install' followed by
a number of 'part --resize' calls. This can leave the board in an
undefined state if any of these calls fails.

This commit introduces the partition variant concept. Now the partition
--install command takes two additional arguments, --variant, which is
optional and default to zero and --part to indicate which block device
the partition table should be installed on.