boot: Overhaul of boot code
This makes the boot code more modular, it's even possible to completley
disable the core boot protocol through Kconfig and insert a custom
version.

The new boot code is split in four different modules
 - boot/boot.c          - Core boot functionallity
 - boot/ab_state.c      - Optional A/B state, rollback logic
 - boot/linux.c         - Linux specific helpers (dtb, ramdisk etc)
 - boot/image_helpers.c - BPAK image helpers

This is also one step in the right direction when it comes to
abstracting the boot image format. In the future it should be possible
to use something other than BPAK files for booting. This is however a
larger change that affects the tooling as well.