mmc_core: Add kconfig to override boot part size
Since punchboot these days actually uses the extcsd to report the
correct size of boot partitions we need a way to be backwards compatible
with existing systems.

This option allows the user to override the boot partition size to
something less than the physical size.

Most users should leave this disabled.