crypto/caam: bugfix: ECDSA r/s alignment
The previous assumption was that the CAAM did not like leading zero
bytes on the R/S integers.

The R/S values are DER encoded 2 complement integers that might have a
leading zero byte if the MSB in the first byte is 1. That bit is the
sign bit and ECDSA integers are always positive.

The CAAM expects the R/S integers to have the following lengths:

 - 32 bytes for secp256r1
 - 48 bytes for secp384r1
 - 66 bytes for secp521r1

This means that sometimes the R/S values can be larger than what the
CAAM expects and in that case the zero byte must be stripped.

If however the encoded integer is shorter it must be prepended with zero
bytes to match the above table.