tools/pbstate: Update pb state struct
This adds the additional "board registers". The idéa is to have a few
board specific registers to control custom board specific logic.

These registers can be accessed in the board bootloader code and through
the pbstate library and CLI.

This commit also moves the struct to the public header so it can be
reused by the bootloader code.