tools/pbstate: Add locking support
This commit introduces flock'ing on the state file. This is needed when
the library is used by more than one process.

Before this change the library would load the state on the init call and
keep the state in memory, any modifications are done on the in memory
state and then written. This will not work when multiple processes are
using the library since it's more or less guaranteed to cause data loss.

Because of how flock must be implemented and because we cant keep an in
memory copy of the state most of the functions have changed and the API
has changed on a few places as well. Mostly because now API functions
must load and check the state on every invocation and that means that
they now must be able to return error codes.