cm/wire: Rename PB_CMD_AUTH_SET_OTP_PASSWORD
And drop the "OTP" part. It _may_ be OTP on some platforms/boards but
the wire format / API should not dictate that.

This function has never been implemented in the CM parser so it's safe
to adjust it.

This still needs an implementation / abstraction in CM before it can be
useful.