ci: cibuildwheel setup
This commit includes fixes to enable builds on:
 - Native windows builds without using msys2/cygwin
 - Macos x86_64 and arm64
 - Linux

Historically the only binary release was on windows with an installer to
simplify driver installation and such. With this commit we instead build
binary python wheels for Python 3.8 -> Python 3.12 on Windows, Linux and
MacOS (x86_64 and arm64) and distribute them through PyPi.

It's now as easy as: pip install punchboot.