drivers/usb/usbd: Add Microsoft OS Descriptors
When a new USB device enumerates on Windows the system will ask for string
descriptor 0xEE, Which is Windows specific, if the string descriptor contains
'MSFT100' it will issue a vendor specific request for a 'Microsoft Compatible ID Feature
Descriptor'.

With these two in place Windows will automatically bind the device to WinUSB and
in turn we can use libusb directly.

Historically this was handled by the windows binary installer that bundeled
libwdi.

It's explained in greater detail here:
    https://github.com/pbatard/libwdi/wiki/WCID-Devices