build: Adjust linker section alignment per arch
On smaller systems the various 4k alignments are pretty expensive. With
this change the alignment is configured per arch and for armv7a, armv8a
it remains the same, 4k, since that is typpicale the page granualrity
for the mmu (and we want it to be 4k).

But for MCU's this is excessive and this commit adjusts the section
alignment to 8 bytes.