tools: punchboot: Allow to close a session explicitly
In complex usecases we may end up with a session staying open.

Add a close() method to enable the use of contextlib.closing() helpers
on the session objects in higher level code, to ensure they are closed
on time.

Previously we relied on garbage collection, which wasn't always running
when needed to close the sessions before trying to open new ones.