drivers/mmc: Expose RPMB block device
This exposes the RPMB block device for reading. There is currently no
support for writing to RPMB devices yet as it requires HMAC. There is
also more missing in the mmc driver before the RPMB can be provisioned
from punchboot.

This is useful if the RPMB is provisioned elsewhere and we only want to
read things from RPMB here.