drivers: mmc: Don't switch back to user part
This interferes with RPBM access. At least on some devices it seems to
reset the internal state between writing a command and reading the
result. Therfore we can't change the partition access like this.

Normally both boot partitions are updated and therefor it should be okay
if either one is set to active.

And regardless of partition access setting there is always one active
boot partition so we should not be able to end up in a non bootable
state as the comment suggested.

The R/O issue fixes it self in the next boot cycle since we're usually
going to be reading from the USER area, which put's the boot R/W
settings back to normal.