drivers: mmc: RPMB support
Access to RPMB always uses the multiple blocks command even if there's
just one block to be handled. RPMB access also requires us to set
the block count. Because the standard says so.

This seems to be different between manufacturers, at least some Micron
memories seem fine without this and just works when reading blocks.
Other memories refuse to do read accesses without this and the standard
also says that it should be done in this way.