cm: add argument to slc_set_eol
On some platforms, such as imx8x, an authenticated message is necessary
to advance the life cycle to EOL. Prepare the EOL SLC API's to receive
such a message.

For imx8x, knowing the length of the message is not strictly necessary
as it will just relay this message to the SCU and eventually the SECO.
For future purposes, a length argument might be necessary. Therefore,
add it right away.