plat: imx8x: add support for reading last seco event
The SECO is able to provide additional events which may contain things
such as extended error codes to previous RPC calls.

It's quite rough on the edges as the documentation states that it has
a fixed number of events it can store and once the maximum number of
events have been reached, it will no longer store any more events.

The event codes are still useful in some cases for debugging purposes.
Add support for reading the last recorded event.